requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
faiss-cpu
sentence-transformers
torch
aiohttp
python-dotenv